Temp fix for question 3 not rendering correctly
There is something fundamentally wrong with how the javascript logic works in JPEG-XL_FAQ.js. It was a group effort, but we found out that bug happens with the toggling of answers to be shown/hidden, something about how nested bullet points don't get unhidden when they should be unhidden. I don't have the time or energy to fix this today, but this will work for now.